requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-decouple==2.1
django-environ==0.9.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
Pillow==9.3.0
psycopg2==2.9.5
psycopg2-binary==2.9.3
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
whitenoise==6.2.0
